+++ {{cookiecutter.repo_name}}/project/webapp_statics/js/jquery/random-item.js
@@ -20,14 +20,6 @@ $.fn.RandomItem = function(method) {
 };
 
 /*
- * Simple image preloader method
- */
-var image_preloader = function(image) {
-    var img = new Image();
-    img.src = image;
-}
-
-/*
 * Plugin extension methods
 */
 var extensions = {